chore: create a projet_service file
