+++ gsl_programmation/tests/services/test_projet_services.py
@@ -8,7 +8,7 @@ from gsl_demarches_simplifiees.tests.factories import (
     NaturePorteurProjetFactory,
 )
 from gsl_programmation.models import Simulation
-from gsl_programmation.services import ProjetService
+from gsl_programmation.service.projet_service import ProjetService
 from gsl_programmation.tests.factories import SimulationFactory, SimulationProjetFactory
 from gsl_projet.models import Projet
 from gsl_projet.tests.factories import ProjetFactory
